back-service/requirements.txt
Flask
Werkzeug
requests
opentracing
opentelemetry-api
opentelemetry-distro
opentelemetry-instrumentation
opentelemetry-exporter-otlp